update project structure / circular imports issue...
