+++ examples/mw_example.py
@@ -11,7 +11,7 @@ from mictlanx.v4.client import Client as MictlanXClient
 from mictlanx.utils.index import Utils as UtilsX
 from xolo.utils.utils import Utils as XoloUtils
 # 
-from axo import Axo,axo_method,axo_task
+from axo import Axo,axo_method
 from axo.contextmanager import AxoContextManager
 from axo.endpoint import EndpointManagerX
 from axo.storage import StorageService
